Features: 1). Added A costume loggger for kivystudio 2). an inline logging feature
