+++ kivystudio/components/terminal/__init__.py
@@ -18,13 +18,14 @@ class TerminalSpace(BoxLayout):
 
     def __init__(self, **k):
         super(TerminalSpace, self).__init__(**k)
-        error_logger = ErrorLogger()
-        self.add_widget(error_logger, title='Logs')
+        self.logger = ErrorLogger()
+        self.add_widget(self.logger, title='Logs')
 
     def add_widget(self, widget, title=''):
         if len(self.children) > 1:
             tab = TerminalTab()
             tab.text=title
+            tab.name=title
             tab.bind(state=self.tab_state)
             self.tab_container.add_widget(tab)
             screen = Screen(name=title)
@@ -35,14 +36,14 @@ class TerminalSpace(BoxLayout):
 
     def tab_state(self, tab, state):
         if state=='down':
-            manager.current = tab.text 
+            self.manager.current = tab.name
 
 class TerminalTab(ToggleButtonBehavior, Label):
 
     def on_state(self, *a):
         if self.state=='down':
             self.text = "[u]" + self.text + "[/u]"
-            self.color = (1,1,1,1)
+            self.color = (.9,.9,.9,1)
         else:
             self.text = self.text.replace('[u]','').replace('[/u]','')
             self.color = (.5,.5,.5,1)
@@ -52,15 +53,32 @@ Builder.load_string('''
     tab_container: tab_container
     manager: manager
     orientation: 'vertical'
-
-    ScrollView:
+    pos_hint: {'y': 0, 'center_x': .5}
+    size_hint_y: .4
+    canvas.before:
+        Color:
+            rgba: .12,.12,.12,1
+        Rectangle:
+            size: self.size
+            pos: self.pos
+        Color:
+            rgba: 1,1,1,1
+        Line:
+            points: [self.x,self.top,self.right,self.top]
+            width: dp(1.4)
+    BoxLayout:
         size_hint_y: None
-        height: '28dp'
+        height: '48dp'
         GridLayout:
             id: tab_container
             rows: 1
+        BoxLayout:
             size_hint_x: None
             width: self.minimum_width
+            IconLabelButton:
+                icon: 'fa-close'
+                size_hint_x: None
+                width: '12dp'
 
     ScreenManager:
         id: manager
@@ -68,6 +86,7 @@ Builder.load_string('''
 <TerminalTab>:
     allow_no_selection: False
     size_hint_x: None
-    width: '180dp'
+    width: '94dp'
+    markup: True
 
 ''')
