+++ kivystudio/main.kv
@@ -1,4 +1,5 @@
 #: import studio_app kivystudio.get_kivystudio_app
+#: import icon kivystudio.tools.iconfonts.icon
 
 <Assembly>:
     orientation: 'vertical'
